+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2025-02-12
+- Added new "Relation to an anime on an AWC Community List" filter for manga
+
 ## 2025-01-09
 - Added new ID filter